+++ src/js/components/agencySubmissionStats/componentMapping/modalContentMapping.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import PublicationDatesContainer from 'containers/aboutTheData/modals/PublicationDatesContainer';
-import MissingAccountBalanceContainer from 'containers/aboutTheData/modals/MissingAccountBalanceContainer';
-import ReportingDifferencesContainer from 'containers/aboutTheData/modals/ReportingDifferencesContainer';
-import UnlinkedDataContainer from 'containers/aboutTheData/modals/UnlinkedDataContainer';
+import PublicationDatesContainer from 'containers/agencySubmissionStats/modals/PublicationDatesContainer';
+import MissingAccountBalanceContainer from 'containers/agencySubmissionStats/modals/MissingAccountBalanceContainer';
+import ReportingDifferencesContainer from 'containers/agencySubmissionStats/modals/ReportingDifferencesContainer';
+import UnlinkedDataContainer from 'containers/agencySubmissionStats/modals/UnlinkedDataContainer';
 
 export const modalContentMapping = (data) => ({
     publicationDates: (<PublicationDatesContainer {...data} />),
